requirements.txt
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
spotipy==2.23.0
transformers==4.35.0
torch==2.1.0